pyproject.toml
[tool.poetry]
name = "ai-trading-system"
version = "0.1.0"
description = "AI-powered crypto trading system"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = "^0.27.0"
pydantic = "^2.6.0"
pydantic-settings = "^2.1.0"
sqlalchemy = "^2.0.25"
asyncpg = "^0.29.0"
redis = "^5.0.1"
websockets = "^12.0"
httpx = "^0.26.0"
python-dotenv = "^1.0.0"
polars = "^0.20.2"
numpy = "^1.26.3"
aiocache = "^0.12.2"
prometheus-client = "^0.19.0"
openai = "^1.10.0"
langgraph = "^0.0.15"
langchain = "^0.1.0"
loguru = "^0.7.2"
alembic = "^1.13.1"
clickhouse-driver = "^0.2.6"
opentelemetry-api = "^1.22.0"
opentelemetry-sdk = "^1.22.0"
opentelemetry-instrumentation-fastapi = "^0.43b0"
opentelemetry-exporter-otlp = "^1.22.0"
sentry-sdk = {extras = ["fastapi"], version = "^1.40.0"}
greenlet = "^3.0.3"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.5"
pytest-cov = "^4.1.0"
black = "^24.1.1"
isort = "^5.13.2"
mypy = "^1.8.0"
ruff = "^0.2.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88

[tool.mypy]
python-version = "3.11"
strict = true

[tool.ruff]
line-length = 88
target-version = "py311" 